+++ CHANGELOG.md
@@ -44,6 +44,7 @@ Add a link to the GitHub diff like
 
 ### Fixed
 - adornment proxies are generated with incorrect relative path
+- execution of Git post-install script
 
 ## [0.8.0] - 2016-02-15
 [Full Changelog](https://github.com/mastersign/bench/compare/v0.7.2...v0.8.0)